Extracted game loop threading into desktop project.
The `GameLoop` is still in the game, just all the threading logic has been moved out. It now only has a `GameLoop.step()` function to advance it to the next frame.

This allows the specific platform more control over how and when steps are scheduled. It also makes it easy to pass input state and graphics canvas for the next step. This also will go a long way in porting it to new platforms.