Moved Rendering and Resource loading out of the game.
This is now done by the specific platform project. The way resources are loaded depends strongly on the platform that is being targeted. As such, the platform now needs to implement the loading behavior. Drawing resources can now be done from a global list of known resources, loading and rendering the actual images is left to the platform to implement.