Build-time task to generate Base64 web resources.
We compile the full list of all known `ImageResources` at build-time, Base64 encode each file and generate a source-file that contains everything as constants. This allows us to ship all resources with the code, so we don't need to wait for additional requests to load resources.

This is a fine approach for us, since our resources are very small PNG images. For larger resources, we'd probably want a more sophisticated loading strategy. This way is simple though, since we don't need to wait (loading screen) or handle load failures in any way.