FIX: Render pacman on web without constant flashing.
In an edge case where the mouth was fully closed, the web renderer would render a full circle instead of nothing (which is what the Swing renderer does). The full circle drawn on top of the body - another full circle - caused a visible flash every few draw calls.

Clamped the value to `0` to avoid this situation in the renderer. Flashes on the web are now gone, all without any double buffering!